[refactor] 카테고리 필드 삭제, baseBoard 변경
