--- src/main/java/com/inhabas/api/domain/board/Category.java
@@ -1,24 +1,0 @@
-package com.inhabas.api.domain.board;
-
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import javax.persistence.*;
-
-/**
- * This is for NormalBoard
- */
-@Entity
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Table(name = "normal_board_category")
-public class Category {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
-
-    private String name;
-
-    private String description;
-}
