+++ src/main/java/com/inhabas/api/domain/board/NormalBoard.java
@@ -1,26 +1,47 @@
 package com.inhabas.api.domain.board;
 
+import com.inhabas.api.domain.BaseEntity;
 import com.inhabas.api.domain.board.type.wrapper.Contents;
 import com.inhabas.api.domain.board.type.wrapper.Title;
+import com.inhabas.api.domain.comment.Comment;
+import com.inhabas.api.domain.file.BoardFile;
 import com.inhabas.api.domain.member.Member;
 import lombok.AccessLevel;
-import lombok.Getter;
 import lombok.NoArgsConstructor;
 import org.springframework.data.jpa.domain.support.AuditingEntityListener;
 
 import javax.persistence.*;
+import java.util.*;
+
 
 @Entity
 @Table(name = "normal_board")
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
 @EntityListeners(AuditingEntityListener.class)
-@DiscriminatorValue("Normal")
-public class NormalBoard extends BaseBoard {
+@Inheritance(strategy = InheritanceType.JOINED)
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@DiscriminatorColumn(name = "TYPE", length = 15)
+public class NormalBoard extends BaseEntity {
 
-    @ManyToOne(fetch = FetchType.EAGER, optional = false)
-    @JoinColumn(name = "category_id", foreignKey = @ForeignKey(name = "normal_board_category_fk"))
-    private Category category;
+    @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
+    protected Integer id;
+
+    @Embedded
+    protected Title title;
+
+    @Embedded
+    protected Contents contents;
+
+    @ManyToOne(fetch = FetchType.LAZY, optional = false)
+    @JoinColumn(name = "writer_id", foreignKey = @ForeignKey(name = "fk_baseboard_to_user"))
+    protected Member writer;
+
+    @OneToMany(mappedBy = "parentBoard", cascade = CascadeType.ALL, orphanRemoval = true)
+    protected List<Comment> comments = new ArrayList<>();
+
+    @OneToMany(mappedBy = "parentBoard", cascade = CascadeType.ALL, orphanRemoval = true)
+    protected Set<BoardFile> files = new HashSet<>();
+
+    /* constructor */
 
     public NormalBoard(Integer id, String title, String contents) {
         this.id = id;
@@ -33,13 +54,57 @@ public class NormalBoard extends BaseBoard {
         this.contents = new Contents(contents);
     }
 
-    @Override
-    public NormalBoard writtenBy(Member writer) {
-        return (NormalBoard) super.writtenBy(writer);
+    /* getter */
+
+    public Integer getId() {
+        return id;
     }
 
-    public NormalBoard inCategoryOf(Category category) {
-        this.category = category;
+    public String getTitle() {
+        return title.getValue();
+    }
+
+    public String getContents() {
+        return contents.getValue();
+    }
+
+    public Member getWriter() {
+        return writer;
+    }
+
+    public Set<BoardFile> getFiles() {
+        return Collections.unmodifiableSet(files);
+    }
+
+    /* relation method */
+
+    public NormalBoard writtenBy(Member writer) {
+        if (Objects.isNull(this.writer))
+            this.writer = writer;
+        else
+            throw new IllegalStateException("게시글 작성자를 수정할 수 없습니다.");
         return this;
     }
+
+    public NormalBoard addFiles(Set<BoardFile> UploadFiles) {
+        if (Objects.nonNull(UploadFiles))
+            UploadFiles.forEach(this::addFile);
+
+        return this;
+    }
+
+    public boolean isWriter(Member member) {
+        return this.writer.equals(member);
+    }
+
+    public void addFile(BoardFile uploadFile) {
+        files.add(uploadFile);
+        uploadFile.toBoard(this);
+    }
+
+    public void addComment(Comment newComment) {
+        comments.add(newComment);
+        newComment.toBoard(this);
+    }
 }
+
