+++ src/main/java/com/inhabas/api/domain/board/NormalBoardRepository.java
@@ -7,9 +7,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
 public interface NormalBoardRepository extends JpaRepository<NormalBoard, Integer> {
-
-    Page<NormalBoard> findAllByCategoryId(Integer categoryId, Pageable pageable);
-
-    List<NormalBoard> findAllByCategoryId(Integer categoryId);
 }
