+++ src/main/java/com/inhabas/api/domain/comment/CustomCommentRepository.java
@@ -1,6 +1,5 @@
 package com.inhabas.api.domain.comment;
 
-import com.inhabas.api.domain.board.BaseBoard;
 import com.inhabas.api.dto.CommentDetailDto;
 
 import java.util.List;
