+++ src/main/java/com/inhabas/api/dto/board/SaveBoardDto.java
@@ -1,7 +1,5 @@
 package com.inhabas.api.dto.board;
 
-import com.inhabas.api.domain.board.Category;
-import com.inhabas.api.domain.member.Member;
 import lombok.Getter;
 import lombok.Setter;
 
@@ -9,11 +7,9 @@ import lombok.Setter;
 public class SaveBoardDto {
     private String title;
     private String contents;
-    private Category category;
 
-    public SaveBoardDto(String title, String contents, Category category) {
+    public SaveBoardDto(String title, String contents) {
         this.title = title;
         this.contents = contents;
-        this.category = category;
     }
 }
