+++ src/main/java/com/inhabas/api/dto/board/UpdateBoardDto.java
@@ -1,6 +1,5 @@
 package com.inhabas.api.dto.board;
 
-import com.inhabas.api.domain.board.Category;
 import lombok.Getter;
 import lombok.Setter;
 
@@ -9,12 +8,10 @@ public class UpdateBoardDto {
     private Integer id;
     private String title;
     private String contents;
-    private Category category;
 
-    public UpdateBoardDto(Integer id, String title, String contents, Category category) {
+    public UpdateBoardDto(Integer id, String title, String contents) {
         this.id = id;
         this.title = title;
         this.contents = contents;
-        this.category = category;
     }
 }
