+++ src/test/java/com/inhabas/api/domain/BaseEntityTest.java
@@ -2,13 +2,11 @@ package com.inhabas.api.domain;
 
 import com.inhabas.api.config.JpaConfig;
 import com.inhabas.api.domain.board.NormalBoard;
-import com.inhabas.api.domain.board.Category;
 import com.inhabas.api.domain.member.Member;
 import com.inhabas.api.domain.board.NormalBoardRepository;
 import com.inhabas.api.domain.member.MemberRepository;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 import org.springframework.context.annotation.Import;
 
@@ -21,7 +19,6 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 @DataJpaTest
 @Import(JpaConfig.class)
-//@AutoConfigureTestDatabase(replace = AutoConfigureTestDatabase.Replace.NONE)
 public class BaseEntityTest {
 
     @Autowired
@@ -36,8 +33,7 @@ public class BaseEntityTest {
         //given
         memberRepository.save(MEMBER1);
         NormalBoard board = new NormalBoard("title", "contents")
-                .writtenBy(MEMBER1)
-                .inCategoryOf(em.getReference(Category.class, 2));
+                .writtenBy(MEMBER1);
 
         //when
         NormalBoard save = boardRepository.save(board);
@@ -52,14 +48,12 @@ public class BaseEntityTest {
         //given
         Member member = memberRepository.save(MEMBER1);
         NormalBoard board = new NormalBoard("title", "contents")
-                .writtenBy(member)
-                .inCategoryOf(em.getReference(Category.class, 2));
+                .writtenBy(member);
         boardRepository.save(board);
 
         //when
         NormalBoard param = new NormalBoard(board.getId(), "new title", "new contents")
-                .writtenBy(member)
-                .inCategoryOf(em.getReference(Category.class, 2));
+                .writtenBy(member);
         NormalBoard updateBoard = boardRepository.save(param);
         em.flush();
