+++ src/test/java/com/inhabas/api/domain/MemberRepositoryTest.java
@@ -22,7 +22,6 @@ import static org.junit.jupiter.api.Assertions.assertAll;
 
 @DataJpaTest
 @Import(JpaConfig.class)
-//@AutoConfigureTestDatabase(replace = AutoConfigureTestDatabase.Replace.NONE)
 public class MemberRepositoryTest {
 
     @Autowired
