--- src/test/resources/data.sql
@@ -1,10 +1,0 @@
-insert into normal_board_category (id, name, description)
-values  (1, '공지사항', '부원이 알아야 할 내용을 공지합니다.'),
-        (2, '자유게시판', '부원이라면 모두 사용할 수 있는 게시판입니다.'),
-        (3, '질문게시판', '어떤 부원이든 질문할 수 있는 게시판입니다.'),
-        (4, '동아리 활동', '동아리 수상 및 활동 내역에 대한 게시판입니다.'),
-        (5, '전체게시판', '모든 게시글이 게시됩니다.'),
-        (6, '알파테스터', 'IBAS 내부 컨테스트, 알파테스터 관련 게시판입니다.'),
-        (7, '베타테스터', 'IBAS 내부 컨테스트, 베타테스터 관련 게시판입니다.'),
-        (8, '회장단 게시판', '회장단 자료 공유 게시판입니다.'),
-        (9, '건의사항', '건의사항이 있다면 말씀해주세요');
